# backend/src/config/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signet Workforce ERP API",
    "version": "1.0.0",
    "description": "Production REST API for Signet Workforce ERP. Authenticate via POST /api/auth/login, then use the Bearer token for protected routes."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "InvoiceLineItem": {
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unitRate"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unitRate": {
            "type": "number",
            "minimum": 0
          },
          "hsnSacCode": {
            "type": "string",
            "example": "998519"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/billing/invoices/site/{siteId}": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create manual invoice for a specific site",
        "parameters": [
          {
            "in": "path",
            "name": "siteId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "invoiceDate",
                  "dueDate",
                  "month",
                  "year",
                  "gstRate",
                  "lineItems"
                ],
                "properties": {
                  "invoiceDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "month": {
                    "type": "integer"
                  },
                  "year": {
                    "type": "integer"
                  },
                  "gstRate": {
                    "type": "number"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "termsAndConditions": {
                    "type": "string"
                  },
                  "lineItems": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/InvoiceLineItem"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created site invoice"
          }
        }
      }
    },
...
    "/api/sites": {
      "get": {
        "tags": [
          "Sites"
        ],
        "summary": "List sites",
        "parameters": [
          {
            "in": "query",
            "name": "clientId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Site list"
          }
        }
      }
    },
...
  },
...
}